Implement basic recurring events functionality
We depend on https://github.com/simshaun/recurr for the recurrence calculation.

Reccurent EventDates are generated after the tca form is saved by the DataHandler,
We generate a database record for all dates described by the recurrent eventdate.

The advantage: Frontend rendering is easy and fast.
We're able to use the extbase query interface without having to generate soft dates.