EventDate::start/end: Switch database type to a simple int
Using MySQL DateTime we had unreliable handling of timezones.
In short: The frontend output was different than what the user entered in the backend.
Backend seems to safe as GMT while frontend respects the server settings (in our case GMT+1)

That needs more investigation and analysis, but for now we opt for the simple
integer-based unix timestamp fields which just work.