Defer recurrence creation until `event` is set in the eventdate table
Unfortunately TYPO3's inline saving is not atomic.
The DataHandler first writes all the entfered data to the database and
sometimes later fills the pointer field (in our case `event`) to the parent.

This means the `event` field is not yet set in the afterDatabaseOperations
(where we create the cloned recurrences). Therefore we defer new records
to the "afterAllOperations" phase.

NOTE: Since the inline handling is not atomic we may need to add a fix
in the frontend that adds some default demands that only select dates
where the `event` pointer is set  (to avoid a possible race condition).