Remove multilingual support for EventDates
EventDates (note the Dates, not the Event itself) contains
only numeric or static data therefore a translation is useless.

Also we'd need to sync the sys_language_uid fields to have proper
multilanguage support. Dropping this is both easiest for our code
and for the editors since they have less to translate.